resolvendo bugs
