+++ src/pages/RestaurantPage/RestaurantPage.js
@@ -1,15 +1,31 @@
-import React from "react"
-import { useNavigate } from "react-router-dom"
+import React, {useContext} from "react"
+import { useNavigate, useParams } from "react-router-dom"
 import Back from '../../assets/back.png'
-import CardFood from "../../components/CardFood/CardFood"
+import Card from "../../components/Card/Card"
+import { BASE_URL } from "../../constants/url"
+import useRequestData from "../../hooks/useRequestData"
 import { goToFeed } from "../../routes/coordinator"
 import CardRestaurant from "./CardRestaurant/CardRestaurant"
-import { Container, BackImg, ContainerFoods, PageTittleContainer, TittleNavContainer } from './styled'
+import { Container, BackImg, ContainerFoods, PageTittleContainer, TittleNavContainer, TitleCategory } from './styled'
+import GlobalStateContext from "../../global/GlobalStateContext";
 
 
 const RestaurantePage = () => {
+  const params = useParams();
+  const {food, setFood} = useContext(GlobalStateContext);
   const navigate = useNavigate();
-  
+  const [foods] = useRequestData([], `${BASE_URL}/restaurants/${params.id}`)
+
+  const openQuantity = (food) => {
+    document.querySelector(`.popup-wrapper`).style.display = `block`
+    setFood(food)
+    console.log(food)
+  }
+
+  const listOfCategories = foods.restaurant && foods.restaurant.products.map((food) => {
+    return food.category
+})
+  const onlyCategories = [...new Set(listOfCategories)]
 
   return (
     <Container>
@@ -23,10 +39,39 @@ const RestaurantePage = () => {
 
       <CardRestaurant />
       <ContainerFoods>
-        <CardFood />
-      </ContainerFoods>
+   {onlyCategories && onlyCategories.map((type) => {
+      return (
+         <div key={type}>
+          <TitleCategory>{type}</TitleCategory>
+          {foods.restaurant && foods.restaurant.products.filter((cat) => {
+              return cat.category === type
+          })
+              .map((food) => {
+                  return (
+                       <Card
+                        key={food.id}
+                        photo={food.photoUrl}
+                        name={food.name}
+                        description={food.description}
+                        price={food.price}
+                        quantityA={food.quantity}
+                        foodAll={food}
+                        id={food.id}
+                        openQuantity={()=>openQuantity(food)}
+                        />
+                  )
+              })}
+      </div>
+  )
+})
+}
+      </ContainerFoods>      
     </Container>
   )
 }
 
-export default RestaurantePage
+export default RestaurantePage;
+
+
+
+
