+++ src/pages/SplashScreen/SplashScreen.js
@@ -1,0 +1,12 @@
+import React from "react"
+
+
+const SplashScreen = () => {
+
+    return(
+        <div>
+
+        </div>
+    )
+}
+
